+++ CHANGELOG.md
@@ -1,13 +1,49 @@
 # go-fil-markets changelog
 
+# go-fil-markets v1.0.7
+
+Seperate pre-commit & post-commit states for deals
+
+### Changelog
+
+- github.com/filecoin-project/go-fil-markets:
+  - docs(CHANGELOG): update for v1.0.6 ([filecoin-project/go-fil-markets#454](https://github.com/filecoin-project/go-fil-markets/pull/454))
+  - Add StorageDealAwaitingPreCommit state (#453) ([filecoin-project/go-fil-markets#453](https://github.com/filecoin-project/go-fil-markets/pull/453))
+
+### Contributors
+
+| Contributor | Commits | Lines ± | Files Changed |
+|-------------|---------|---------|---------------|
+| dirkmc | 1 | +349/-37 | 19 |
+
+# go-fil-markets v1.0.6
+
+Complete removal of temp file usage
+
+### Changelog
+
+- github.com/filecoin-project/go-fil-markets:
+  - Integrate go-commp-utils (#452) ([filecoin-project/go-fil-markets#452](https://github.com/filecoin-project/go-fil-markets/pull/452))
+  - release: v1.0.5 ([filecoin-project/go-fil-markets#451](https://github.com/filecoin-project/go-fil-markets/pull/451))
+
+### Contributors
+
+| Contributor | Commits | Lines ± | Files Changed |
+|-------------|---------|---------|---------------|
+| dirkmc | 1 | +51/-854 | 21 |
+
 # go-fil-markets v1.0.5
+
+Minor bug fix release
+
+### Changelog
 
 - github.com/filecoin-project/go-fil-markets:
   - Remove temp files (#449) ([filecoin-project/go-fil-markets#449](https://github.com/filecoin-project/go-fil-markets/pull/449))
   - fix: fail fast in tests that use a waitgroup Wait() (#450) ([filecoin-project/go-fil-markets#450](https://github.com/filecoin-project/go-fil-markets/pull/450))
   - Make it possible to configure the proof type at runtime. ([filecoin-project/go-fil-markets#448](https://github.com/filecoin-project/go-fil-markets/pull/448))
 
-Contributors
+### Contributors
 
 | Contributor | Commits | Lines ± | Files Changed |
 |-------------|---------|---------|---------------|